pyproject.toml
[project]
name = "juniorpen01s-snake"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "juniorpen01", email = "01juniorpen@gmail.com" }]
requires-python = ">=3.13"
dependencies = ["pygame-ce>=2.5.3"]

[project.scripts]
juniorpen01s-snake = "juniorpen01s_snake:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
